Use `updateRepoVersion` from versionutil to update repo's version properly
