+++ src/versionutil.js
@@ -16,9 +16,15 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-var repoutil = require('./repoutil');
-var flagutil = require('./flagutil');
+
+var fs = require('fs');
 var path = require('path');
+var shelljs = require('shelljs');
+var apputil = require('./apputil');
+var executil = require('./executil');
+var flagutil = require('./flagutil');
+var gitutil = require('./gitutil');
+var repoutil = require('./repoutil');
 
 function removeDev(version) {
     var newVersion = version.replace('-dev', '');
@@ -27,12 +33,12 @@ function removeDev(version) {
 exports.removeDev = removeDev;
 
 //updates platformsConfig.json
-//Needs to be passed a object which includes repo.id as key 
+//Needs to be passed a object which includes repo.id as key
 //and the new version as value
 //ex {android:4.0.0}
 function *updatePlatformsConfig(newValues) {
-    
-    var platformsConfig = path.join(repoutil.getRepoDir('cordova-lib'), 
+
+    var platformsConfig = path.join(repoutil.getRepoDir('cordova-lib'),
         'src/cordova/platformsConfig.json');
     console.log(platformsConfig);
     var platformsJS = require(platformsConfig);
@@ -42,11 +48,11 @@ function *updatePlatformsConfig(newValues) {
     yield repoutil.forEachRepo(repos, function*(repo) {
         if(repo.id === 'windows') {
             platformsJS[repo.id].version = newValues[repo.id];
-            platformsJS['windows8'].version = newValues[repo.id]; 
+            platformsJS['windows8'].version = newValues[repo.id];
         } else if(repo.id === 'blackberry') {
-            platformsJS['blackberry10'].version = newValues[repo.id]; 
+            platformsJS['blackberry10'].version = newValues[repo.id];
         } else {
-            platformsJS[repo.id].version = newValues[repo.id]; 
+            platformsJS[repo.id].version = newValues[repo.id];
         }
     });
 
@@ -55,3 +61,67 @@ function *updatePlatformsConfig(newValues) {
     });
 }
 exports.updatePlatformsConfig = updatePlatformsConfig;
+
+/**
+ * Updates VERSION file, version executable script and package.json using specified
+ *   version. Also commits change made to the repo if opposite is not specified.
+ *
+ * @param {Object}  repo    Repo to update version for
+ * @param {String}  version A semver-compatible version to write to repo's files
+ * @param {Object}  [opts]  An options object
+ * @param {Boolean} [opts.commitChanges=true] Specifies whether to commit changes
+ *   to the repo after update is done.
+ */
+exports.updateRepoVersion = function *updateRepoVersion(repo, version, opts) {
+    // Update the VERSION files.
+    var versionFilePaths = repo.versionFilePaths || ['VERSION'];
+    var isPlatformRepo = !!repoutil.isInRepoGroup(repo, 'platform');
+    if (isPlatformRepo && fs.existsSync(versionFilePaths[0])) {
+        versionFilePaths.forEach(function(versionFilePath) {
+            fs.writeFileSync(versionFilePath, version + '\n');
+        });
+        shelljs.config.fatal = true;
+        if (repo.id == 'android' || repo.id == 'amazon-fireos') {
+            shelljs.sed('-i', /CORDOVA_VERSION.*=.*;/, 'CORDOVA_VERSION = "' + version + '";', path.join('framework', 'src', 'org', 'apache', 'cordova', 'CordovaWebView.java'));
+            shelljs.sed('-i', /VERSION.*=.*;/, 'VERSION = "' + version + '";', path.join('bin', 'templates', 'cordova', 'version'));
+        } else if (repo.id == 'ios' || repo.id == 'osx') {
+            shelljs.sed('-i', /VERSION.*=.*/, 'VERSION="' + version + '";', path.join('bin', 'templates', 'scripts', 'cordova', 'version'));
+        } else if (repo.id == 'blackberry') {
+            shelljs.sed('-i', /VERSION.*=.*;/, 'VERSION = "' + version + '";', path.join('bin', 'templates', 'project','cordova', 'lib', 'version.js'));
+        } else if (repo.id == 'firefoxos' || repo.id == 'browser' || repo.id == 'ubuntu') {
+            shelljs.sed('-i', /VERSION.*=.*;/, 'VERSION = "' + version + '";', path.join('bin', 'templates', 'project','cordova', 'version'));
+        } else if (repo.id == 'windows') {
+            if(fs.existsSync(path.join('template', 'cordova', 'version'))) {
+                console.log('version exists');
+                shelljs.sed('-i', /VERSION.*=.*;/, 'VERSION = "' + version + '";', path.join('template', 'cordova', 'version'));
+            }
+        }
+        shelljs.config.fatal = false;
+        if (!(yield gitutil.pendingChangesExist())) {
+            apputil.print('VERSION file was already up-to-date.');
+        }
+    } else {
+        if (isPlatformRepo) console.warn('No VERSION file exists in repo ' + repo.repoName);
+    }
+
+    // Update the package.json VERSION.
+    var packageFilePaths = repo.packageFilePaths || ['package.json'];
+    if (fs.existsSync(packageFilePaths[0])) {
+        fs.readFile(packageFilePaths[0], {encoding: 'utf-8'}, function (err, data) {
+            if (err) throw err;
+            var packageJSON = JSON.parse(data);
+            packageJSON.version = version;
+            fs.writeFileSync(packageFilePaths[0], JSON.stringify(packageJSON, null, "    "));
+        });
+        if (!(yield gitutil.pendingChangesExist())) {
+            apputil.print('package.json file was already up-to-date.');
+        }
+    } else {
+        console.warn('No package.json file exists in repo ' + repo.repoName);
+    }
+
+    var commitChanges = !!(opts ? opts.commitChanges : true);
+    if (commitChanges && (yield gitutil.pendingChangesExist())) {
+        yield executil.execHelper(executil.ARGS('git commit -am', 'Set VERSION to ' + version + ' (via coho)'));
+    }
+}
